requirements.txt
asgiref==3.3.1
attrs==20.3.0
Django==3.1.4
djangorestframework==3.12.2
drf-nested-routers==0.92.5
iniconfig==1.1.1
packaging==20.8
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.1
pytest-django==4.1.0
python-dotenv==0.15.0
pytz==2020.4
sqlparse==0.4.1
toml==0.10.2
